requirements.txt
groq>=0.4.2
keras>=2.14.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytz>=2023.3
tensorflow>=2.14.0
joblib>=1.3.0
yfinance>=0.2.33
schedule>=1.2.1
feedparser>=6.0.10
beautifulsoup4>=4.12.0
flask>=3.0.0
gunicorn>=21.2.0
